created controller and added code
