+++ src/modules/workspace/controllers/ai-assistant.controller.ts
@@ -1,5 +1,4 @@
-import { Body, Controller, Post, Req, Res, UseGuards, Get, Query, Delete } from "@nestjs/common";
-import { ApiQuery } from '@nestjs/swagger';
+import { Body, Controller, Post, Req, Res, UseGuards } from "@nestjs/common";
 import { AiAssistantService } from "../services/ai-assistant.service";
 import { FastifyReply } from "fastify";
 import { HttpStatusCode } from "@src/modules/common/enum/httpStatusCode.enum";
@@ -16,8 +15,6 @@ import {
   ErrorResponsePayload,
   ChatBotPayload,
 } from "../payloads/ai-assistant.payload";
-import { LlmConversation } from "../payloads/llm-conversation.payload";
-import { LlmConversationService } from "../services/llm-conversation.service";
 import { ExtendedFastifyRequest } from "@src/types/fastify";
 
 @ApiBearerAuth()
@@ -30,7 +27,7 @@ export class AiAssistantController {
    * @param aiAssistantService - Injected AiAssistantService to handle business logic.
    * * @param llmConversationService - Injected LlmConversationService to handle LLM conversation logic.
    */
-  constructor(private readonly aiAssistantService: AiAssistantService , private readonly llmConversationService: LlmConversationService) {}
+  constructor(private readonly aiAssistantService: AiAssistantService ) {}
 
   @ApiOperation({
     summary: "Get a respose for AI assistant",
@@ -78,65 +75,6 @@ export class AiAssistantController {
       "Prompt Generated Successfully",
       HttpStatusCode.CREATED,
       data,
-    );
-    return res.status(response.httpStatusCode).send(response);
-  }
-
-  @Get("get-conversation")
-  @ApiQuery({ name: 'provider', required: true })
-  @ApiQuery({ name: 'apiKey', required: true })
-  @ApiQuery({ name: 'id', required: false })
-  async GetConversation(
-    @Query("provider") provider: string,
-    @Query("apiKey") apiKey: string,
-    @Query("id") id: string,
-    @Res() res: FastifyReply,
-  ) {
-    const data = await this.llmConversationService.getConversation(provider, apiKey, id);
-    const response = new ApiResponseService(
-      "Conversation fetched successfully",
-      HttpStatusCode.OK,
-      data,
-    );
-    return res.status(response.httpStatusCode).send(response);
-  }
-
-  @Post("insert-conversation")
-  async InsertConversation(@Body() payload: LlmConversation, @Res() res: FastifyReply) {
-    const data = await this.llmConversationService.insertConversation(payload);
-    const response = new ApiResponseService(
-      "Conversation Inserted Successfully",
-      HttpStatusCode.CREATED,
-      data,
-    );
-    return res.status(response.httpStatusCode).send(response);
-  }
-  @Post("update-conversation")
-  async UpdateConversation(@Body() payload: LlmConversation, @Res() res: FastifyReply) {
-    const data = await this.llmConversationService.updateConversation(payload);
-    const response = new ApiResponseService(
-      "Conversation Updated Successfully",
-      HttpStatusCode.CREATED,
-      data,
-    );
-    return res.status(response.httpStatusCode).send(response);
-  }
-
-  @Delete("delete-conversation")
-  @ApiQuery({ name: 'provider', required: true })
-  @ApiQuery({ name: 'apiKey', required: true })
-  @ApiQuery({ name: 'id', required: true })
-  async DeleteConversation(
-    @Query("provider") provider: string,
-    @Query("apiKey") apiKey: string,
-    @Query("id") id: string,
-    @Res() res: FastifyReply,
-  ) {
-    await this.llmConversationService.deleteConversation(provider, apiKey, id);
-    const response = new ApiResponseService(
-      "Conversation deleted successfully",
-      HttpStatusCode.OK,
-      null,
     );
     return res.status(response.httpStatusCode).send(response);
   }
